Added slowdown and scaling and fixed gif creation
Video is scaled to 70% of original size so that it can fit inside
the screen better.

Added a few sleeps to slow down the video so it's easier to spot
mistakes.

Prior to this commit the gifs were stacked on top of eachother
(e.g. 1st gif was good, 2nd gif was 1st + frames that should be in
the second one and so on). This was fixed in this commit.